Added LoadWindow, Results File, Performance Improvements
Added a loading window which opens every time during the start of the program.
All the input excel data files now load only once during the start of the program (instead of every time a new id is checked for eligibility) in the background of loading window.
While checking the eligibility of multiple students from of a list of IDs, the results are now saved in a excel file at two locations-
at the current working directory and also at one directory up the current directory. This is so that user can easily locate and view the results excel file.
While cheking ID, it now checks for the range 0000 to 2500.