Added Distribution Folder, Modified README.md
An executable version of the software is also inlcuded in the 'PS_Eligibility_Software (Distribution)' folder. It was created using pyinstaller. The folder contains a windows batch file and a text file with instructions to use the software. The batch file is used to run the software so that users need not search for the .exe file inside 'PS_Eligibility_Software' directory which makes it easier to use the software and prevents accidental deletion of any file inside that directory by the user. The 'Results_PS1/PS2.xls' is created at two locations, at the same directory as the batch file and inside the 'PS_Eligibility_Software' folder, so that users need not search for the file inside the 'PS_Eligibility_Software' folder. This is also done to prevent accidental deletion of any file inside that directory by the user.